+++ importaciones/WEB-INF/classes/modules/import/classes/om/BaseClientQuotationItem.php
@@ -38,7 +38,7 @@ abstract class BaseClientQuotationItem extends BaseObject  implements Persistent
 
 	/**
 	 * The value for the price field.
-	 * @var        int
+	 * @var        double
 	 */
 	protected $price;
 
@@ -145,7 +145,7 @@ abstract class BaseClientQuotationItem extends BaseObject  implements Persistent
 	/**
 	 * Get the [price] column value.
 	 * precio de producto
-	 * @return     int
+	 * @return     double
 	 */
 	public function getPrice()
 	{
@@ -233,13 +233,13 @@ abstract class BaseClientQuotationItem extends BaseObject  implements Persistent
 	/**
 	 * Set the value of [price] column.
 	 * precio de producto
-	 * @param      int $v new value
+	 * @param      double $v new value
 	 * @return     ClientQuotationItem The current object (for fluent API support)
 	 */
 	public function setPrice($v)
 	{
 		if ($v !== null) {
-			$v = (int) $v;
+			$v = (double) $v;
 		}
 
 		if ($this->price !== $v) {
@@ -310,7 +310,7 @@ abstract class BaseClientQuotationItem extends BaseObject  implements Persistent
 			$this->id = ($row[$startcol + 0] !== null) ? (int) $row[$startcol + 0] : null;
 			$this->clientquotationid = ($row[$startcol + 1] !== null) ? (int) $row[$startcol + 1] : null;
 			$this->productid = ($row[$startcol + 2] !== null) ? (int) $row[$startcol + 2] : null;
-			$this->price = ($row[$startcol + 3] !== null) ? (int) $row[$startcol + 3] : null;
+			$this->price = ($row[$startcol + 3] !== null) ? (double) $row[$startcol + 3] : null;
 			$this->quantity = ($row[$startcol + 4] !== null) ? (int) $row[$startcol + 4] : null;
 			$this->resetModified();
